requirements.txt
azure_storage==0.37.0
datasets==2.19.0
fitz==0.0.1.dev2
llama_index==0.12.8
matplotlib==3.8.4
numpy==2.2.1
openai==1.58.1
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0
tiktoken==0.7.0
tqdm==4.66.4
